Fix multi argument methods
Given the expression `add 4 5`, in parsing the first argument for `add`, we will
put both `4` and `5` in the values in order to check whether they belong to some
other high-precedence function.

In the end, after knowing that `4` and `5` do not belong to other functions, we
should only return `4` and put `5` back to the unparsed words.