Change semantics of optional context params
Two motivations:

- Avoid interference with control of context parameters at the top-level
- Avoid it is being used to introduce global state

The 2nd is subtle ---- of course the architect can control it at the top-level.
However, he cannot control the submodule from changing an immutable default
context to a mutable.

With the semantic change, a mutable global state has to be explicitly bound at
top-level. It cannot be implicitly bound at submodule-level with only a single
`allow` at the top-level.