Improve exhaustivity check
This change is motivated by the following case

    match cons case Len n =>

In the above, `Len` matchs `List[Int]`, however the scrutinee is `Cons[Int]`.

However, it has the negative consequence that if the match `Len` is not
exhaustive and later users provide a pattern `Cons[Int]`, an inexhaustive
warning will be incorrectly produces.

In practice, it is not expected to happen: either the scrutinee is `List[Int]`
or programmers make `Len` exhaustive.