Change semantics of optional context parameters
The previous semantics incur hassle in implementation due to the awkward
encoding via another context parameters and insertion of None bindings in
many places.

One would want to do a direct implementation by checking the presence of the
binding for the optional context parameter. That does not work. We also need
to make sure that all bindings for an optional context parameter disappear
when going into a effect boundary where they are not allowed!

In the end, it seems to me a good compromise is to bind default context
parameters at effect boundaries where they are not allowed. This way, we can
ensure that default context parameters are always present.

However, the implementation may have the side effect that if the default value
contains mutable state, that state will be shared in the binding scope. This
may cause uneasiness. However, it is not a big problem in practice:

1. The default values are mostly stateless

2. Top-level programs can always use "allow" to make sure every call to a
   submodule will create a fresh copy of the state, which the submodule can
   effectively do so if they want.

3. The submodule may not make any assumptions on the state given that it is
   parameterizable.

So, in the end, the simplicity of implementation dominates here.